particle time uses double (#2434)
* particle time uses double

* add comments

* multiplier should use single precision float!

* fix typo

* cleaner data types

* minor

+++ include/vapor/AdvectionIO.h
@@ -25,7 +25,7 @@ auto OutputFlowlinesNumSteps(const Advection *adv,
 // Otherwise, only trajectories are output.
 auto OutputFlowlinesMaxTime(const Advection *adv,
                             const char *filename,
-                            float maxTime,
+                            double maxTime,
                             const std::string &proj4string,
                             bool append) -> int;
